Don't build nested kustomize dirs
There is no point to build the nested kustomize directories
since those are not going to get deployed (argo isn't configured
to recursively deploy manifests).

In addition, it can create duplicate manifests, since they will get
built once by the root kustomization and second in the nested
kustomization.

Add a test to verify that nested directories are not getting built.

This change doesn't mean that manifests in the tenant's dir can't
be organized in sub directories. This is still an option, but those
manifests would need to be referenced in the root kustomization file
of the tenant.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>