Do not pin EC policy data sources
As learned in https://issues.redhat.com/browse/RHTAPBUGS-856, pinning
policy data sources can lead to violations due to stale data. This
commit ensures the lates policy data sources are used.

It is still recommended to pin the policy rule sources as these refer to
rego code that needs to be properly gated before deploying. However,
this process can be cumbersome. https://issues.redhat.com/browse/EC-193
aims to address this for the EC policies in this repo.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>