Warn only for centos-bootc high severity CVEs
A followup from #133. The CVEs being detected require upstream fixes
and aren't likely to be fixed in the short term.

Rather than skipping the CVE check entirely, we can provide
customized rule data so the check will not produce a violation
unless critical severity CVEs are found.

This patch can be removed in future when/if the high severity CVEs
are fixed.