refactor(RHTAPREL-890): warn for high sev CVEs in release-service
There is a CVE in ubi9
https://access.redhat.com/security/cve/CVE-2024-27316. This commit
modifies the EnterpriseContractPolicy that the release-service team
applications use (release-service, internal-services) to warn for high
severity CVEs instead of restrict. Once STONEINTG-877 is complete,
things won't fail if a CVE doesn't have a fix yet, like this one. A
comment is added saying the same.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>