Stop using deprecated EC pipelines
In the build-definitions repo, there are 5 different enterprise-contract
pipelines. In a previous generation of Konflux, it was not possible to
set Pipeline parameters in an IntegrationTestScenario. For this reason,
copies of the enterprise-contract Pipeline were introduced. The only
difference between them being the default value of the
`POLICY_CONFIGURATION` parameter.

As captured in EC-330, we want to undo this workaround for various
reasons. The most pressing one being that having to test the different
variations of the Pipeline is time consuming and significantly slows
down e2e tests.

This commit changes all the IntegrationTestScenarios to use the main
enterprise-contract Pipeline, `enterprise-contract.yaml`. In most cases,
the `POLICY_CONFIGURATION` parameter was already overwritten. Simply
changing the Pipeline should yield the same result. In the other cases,
the `POLICY_CONFIGURATION` parameter is now added with the corresponding
value.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>