Create a second SA for konflux-ci releases
There is some bug in `cosign` such that if two secrets are both linked
to push to quay, cosign doesn't know how to sort things out and use the
right one.

To workaround, we need two different serviceaccounts here: one for
pushing to quay.io/redhat-appstudio and one for pushing to
quay.io/konflux-ci.

Signed-off-by: Ralph Bean <rbean@redhat.com>