Run build-service pipelines in rhtap-build-tenant (#330)
STONEBLD-1937

We're onboarding build-service to Konflux, but build-service already has
Tekton pipelines running in the tekton-ci namespace. Because of that,
build-service cannot be onboarded to the rhtap-build-tenant namespace
(see https://pipelinesascode.com/docs/guide/repositorycrd/ ):

* the Repository is already present in the tekton-ci namespace
* the Repository must only exist in one namespace
* the pipelines must run in the same namespace

To avoid losing CI completely for an extended period of time, we want to
keep the old pipelines working while we slowly onboard to Konflux. That
means we need to move the old pipelines from the tekton-ci namespace to
the rhtap-build-tenant namespace.

To do that, we need to:

* delete Repo from tekton-ci (in redhat-appstudio/infra-deployments)
* add Repo to rhtap-build-tenant (Konflux will do that automatically)
* add quay-push-secret (for quay.io/redhat-appstudio/build-service)

Signed-off-by: Adam Cmiel <acmiel@redhat.com>