Update trusted-artifacts application
The application `trusted-artifacts` was previously created with the name
`build-trusted-artifacts`. However, once it was re-onboarded to Konflux,
the new name was chosen.

This commit changes the relevant resources to use the new application
name, with the matching new component name.

It also removes the ReleasePlanAdmission resources as that is now
defined in a different repository.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>