Remove release plan for ec-cli v0.1-alpha
We're no longer maintaining the release-v0.1-alpha branch. It's been
failing the EC check and also its own CI for a while, and it's no
longer listed in the RPA so we're not pushing the builds anywhere.

Let's begin to retire it officially by deleting the release plan. In
future we'll probably also remove the Konflux app.

Ref: https://issues.redhat.com/browse/EC-482